+++ examples/minimal.rs
@@ -5,5 +5,5 @@ extern crate log;
 fn main() {
     fastlog::LogBuilder::new().build().unwrap().init().unwrap();
     info!("Hello, world.");
-    log::shutdown_logger().unwrap();
+    log::logger().flush();
 }